Version 0.0.1.7
Included NOR, XOR and NAND functions, you could argue there's no need for these to be functions as they're so simple, but I thought as it's something I'm sure everyone uses/implements all the time, may as well implement them as functions here. Additionally, they're not compressed at all, this is purely to ensure that the readability is still maintained quite nicely.